Fix diffIn* use across the project
In Laravel 11 the diffIn* carbon methods started to return
positive/negative values based on the direction of the diff.
In order to maintain the expected behavior, we need to ensure
the order of diff is always from the older to the newer date.

If the order is from newer to older, diff would be a negative
number.