Fix failing migrations if mysql client is not present
Beam and Campaign now publish their database schema
to speed up migrations and avoid breaking changes
in the past migrations introduced in Laravel 11/12.
Column updates don't preserve untouched attributes
but remove them unless the update explicitly states it.

However if the migration file exists, Laravel doesn't
check if there's a client that it can use to import
it or not, it just executes it. If the client is not
present, migrations fail.

This changes publishes the schema migration only if
the client is present.

Any database issues caused by applying old migrations
which work differently now will be addressed on
case-by-case bases.

remp/remp#1413