Fix migrations after Laravel's change of migration behavior
Laravel used to touch only the parameter of column that we
were altering within the migration. Now we always have to
state all the parameters of the changed column.

Since we cannot rely that people will be able to use SQL
dump to make the initial import, we needed to make sure
that the old migrations are in sync with new ones.

remp/remp#1409