Fix pageviews count in /api/v2/articles/top if param `to` is provided
Fixed condition (which decides which pageviews count source to use) to
incorporate a datetime parameter `to`.

Possible sources:

- `articles.pageviews_count` -> total count of pageviews,
- `article_pageviews` -> hourly aggregation of pageview counts.

Before fix, we would return total pageviews (if `published_from` was
used and it was bigger than `from`), ignoring `to` parameter.

Now, if `to` parameter is provided, we check if against current date and
if it is smaller, we have to use table with hourly data to return
correct count.

remp/respekt#441